fix: cap tag size to prevent overflow
