+++ src/components/Tag/styles.css
@@ -1,8 +1,3 @@
-.NLT__tag-content {
-    width: max-content;
-    user-select: none;
-}
-
 .NLT__tag {
     display: flex;
     align-items: center;
