+++ src/components/TagCellEdit/styles.css
@@ -2,16 +2,6 @@
 	background-color: var(--background-primary);
 }
 
-.NLT__tag-menu-header {
-	background-color: rgba(0, 0, 0, 0.25);
-	padding: 4px 10px;
-}
-
-.NLT__tag-menu-body {
-	min-height: 100px;
-	padding: 4px 10px;
-}
-
 .NLT__tag-menu-text {
 	font-size: 0.85rem;
 	margin: 5px 0px;
